First Draft of the sprint announcement
needs adding -- what we want to do at the sprint and whether it is
at my house or at OE.  Note that I am fine with you eating and hacking
at my house, it is the sleeping over that is not working due to the
lack of showers.  I suppose it is possible that somebody will want to
begin work on the bathroom building that week, in which case we would
have to have the sprint at OE on the grounds of it being too noisy and
dirty at our place, but I think the chances of this is small.